fix: revert to supported react/dom versions. Support enzyme shallow rendering
