+++ App.test.js
@@ -1,9 +1,9 @@
+import './utils/test.config.js'
 import React from 'react';
-import App from './App';
 import renderer from 'react-test-renderer';
-import { shallow, mount, render } from 'enzyme'
+import { shallow, render } from 'enzyme'
 import { View } from 'react-native'
-import { setup, silenceConsoleError } from './test.config.js'
+import App from './App';
 
 describe(`App`, () => {
   it('renders without crashing', () => {
@@ -13,17 +13,9 @@ describe(`App`, () => {
   
   it(`should shallow render`, () => {
     const wrapper = shallow(<App />)
+    expect(wrapper.state().a).toEqual(1)    
     wrapper.instance().incrementA()
-    expect(wrapper).toBeDefined()
+    expect(wrapper.state().a).toEqual(2)
     expect(wrapper.find(View)).toHaveLength(1)
-  })
-  
-  it(`should mount`, () => {
-    silenceConsoleError(true)
-    const wrapper = mount(<App />)
-    wrapper.instance().incrementA()
-    expect(wrapper).toBeDefined()
-    expect(wrapper.state()).toBeDefined()
-    silenceConsoleError(false)
   })
 })
