+++ package.json
@@ -5,7 +5,7 @@
   "devDependencies": {
     "jest-expo": "^21.0.2",
     "react-native-scripts": "1.5.0",
-    "react-test-renderer": "16.0.0"
+    "react-test-renderer": "16.0.0-alpha.12"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
@@ -24,9 +24,11 @@
     "enzyme-adapter-react-16": "1.0.1",
     "expo": "^21.0.0",
     "jsdom": "11.3.0",
-    "react": "16.0.0",
-    "react-dom": "16.0.0",
+    "jsdom-global": "3.0.2",
+    "react": "16.0.0-alpha.12",
+    "react-dom": "16.0.0-alpha.12",
     "react-native": "^0.48.4",
+    "react-native-mock": "0.3.1",
     "react-native-mock-render": "0.0.9"
   }
 }
